+++ evals/solvers/providers/openai/realtime_solver.py
@@ -10,14 +10,8 @@ import numpy as np
 import websockets
 
 from evals.solvers.solver import Solver, SolverResult
+from evals.solvers.utils import _data_url_to_wav
 from evals.task_state import TaskState
-
-
-def _data_url_to_wav(url):
-    if not url.startswith("data:"):
-        raise ValueError("Not a data URL")
-    _, raw_data = url.split(",", 1)
-    return base64.b64decode(raw_data)
 
 
 def _wav_to_24k_pcm(wav_bytes):
